Controlplane API to include only control plane info
