+++ sensor/controlplane.go
@@ -49,10 +49,6 @@ type ControlPlaneInfo struct {
 	AdminConfigFile       *ds.FileInfo    `json:"adminConfigFile,omitempty"`
 	PKIDIr                *ds.FileInfo    `json:"PKIDir,omitempty"`
 	PKIFiles              []*ds.FileInfo  `json:"PKIFiles,omitempty"`
-	CNIConfigFiles        []*ds.FileInfo  `json:"CNIConfigFiles,omitempty"`
-
-	// The name of the running CNI
-	CNIName string `json:"CNIName,omitempty"`
 }
 
 // K8sProcessInfo holds information about a k8s process
@@ -306,18 +302,6 @@ func SenseControlPlaneInfo() (*ControlPlaneInfo, error) {
 		)
 	}
 
-	// make cni config files
-	CNIConfigInfo, err := makeCNIConfigFilesInfo()
-
-	if err != nil {
-		zap.L().Error("SenseControlPlaneInfo", zap.Error(err))
-	} else {
-		ret.CNIConfigFiles = CNIConfigInfo
-	}
-
-	// get CNI name
-	ret.CNIName = getCNIName()
-
 	// If wasn't able to find any data - this is not a control plane
 	if ret.APIServerInfo == nil &&
 		ret.ControllerManagerInfo == nil &&
@@ -333,68 +317,4 @@ func SenseControlPlaneInfo() (*ControlPlaneInfo, error) {
 	}
 
 	return &ret, nil
-}
-
-// makeCNIConfigFilesInfo - returns a list of FileInfos of cni config files.
-func makeCNIConfigFilesInfo() ([]*ds.FileInfo, error) {
-	// *** Start handling CNI Files
-	kubeletProc, err := LocateKubeletProcess()
-	if err != nil {
-		return nil, err
-	}
-
-	CNIConfigDir := utils.GetCNIConfigPath(kubeletProc)
-
-	if CNIConfigDir == "" {
-		return nil, fmt.Errorf("no CNI Config dir found in getCNIConfigPath")
-	}
-
-	//Getting CNI config files
-	CNIConfigInfo, err := makeHostDirFilesInfoVerbose(CNIConfigDir, true, nil, 0)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to makeHostDirFilesInfo for CNIConfigDir %s: %w", CNIConfigDir, err)
-	}
-
-	if len(CNIConfigInfo) == 0 {
-		zap.L().Debug("SenseControlPlaneInfo - no cni config files were found.",
-			zap.String("path", CNIConfigDir))
-	}
-
-	return CNIConfigInfo, nil
-}
-
-// getCNIName - looking for CNI process and return CNI name, or empty if not found.
-func getCNIName() string {
-	supportedCNIs := []struct {
-		name          string
-		processSuffix string
-	}{
-		// 'canal' CNI "sets up Calico to handle policy management and Flannel to manage the network itself". Therefore we will first
-		// check "calico" (which supports network policies and indicates for either 'canal' or 'calico') and then flannel.
-		{"Calico", "calico-node"},
-		{"Flannel", "flanneld"},
-		{"Cilium", "cilium-agent"},
-		{"WeaveNet", "weave-net"},
-	}
-
-	for _, cni := range supportedCNIs {
-		p, err := utils.LocateProcessByExecSuffix(cni.processSuffix)
-
-		if p != nil {
-			zap.L().Debug("CNI process found", zap.String("name", cni.name))
-			return cni.name
-		}
-
-		if err != nil {
-			zap.L().Error("getCNIName- Failed to locate process for cni",
-				zap.String("cni name", cni.name),
-				zap.Error(err))
-		}
-
-	}
-
-	zap.L().Debug("No supported CNI process was found")
-
-	return ""
 }
